Fix mobile tap logic
I was fighting with this for so long. Mobile "mouse emulation" was not fun to deal with, since it caused double-firing on the onDragEnd call. This means timeslots would toggle on and back off in one tap.

The solution was just to disable mouse events if a tap is detected.