Rewrite painting logic
This includes the addition of shift-painting, which allows you to hold shift to simulate another drag after toggling. Perfect for trackpad users.

Yes, the code is VERY ugly... despite it being a rewrite...

I'll clean it up soon. I'm just so happy I got it to work after wrestling with React states and refs for 3 hours.