Add account button to header with API function
If the user is not logged in, the account button will be a "sign in" button. This is checked against the actual API.

This is not checked again on any page change, but it might be helpful to add a check after a sign in or log out to update it.

The theme toggle button style was also adjusted to play nicely with the container its now in.