Create frosted-glass CSS class
Just to centrally manage that one style.

There is a child selector class for nested frosted-glass elements, so that even in light mode it always has a light background. This is to make sure that any layered frosted glass layers are visible and don't get too dark on top of one another.