Change account button style to frosted glass
It looked weird to have both the new event and log in buttons be red around the frosted glass theme toggle.

Also, the log in button has an annoying 1.5 padding instead of 2, just to keep the size consistent with the theme toggle and account buttons. The joys of using outlines.